Make nonhomology the default
