+++ tardis/model/geometry/radial1d.py
@@ -2,20 +2,13 @@ import warnings
 
 import numpy as np
 from astropy import units as u
-from numba import float64
+from numba import boolean, float64
 from numba.experimental import jitclass
 
 
-class HomologousRadial1DGeometry:
+class Radial1DGeometry:
     """
-    Holds information about model geometry for radial 1D models.
-
-    Parameters
-    ----------
-    v_inner : astropy.units.quantity.Quantity
-    v_outer : astropy.units.quantity.Quantity
-    v_inner_boundary : astropy.units.quantity.Quantity
-    v_outer_boundary : astropy.units.quantity.Quantity
+    Holds information about model geometry for non-homologous radial 1D models.
 
     Attributes
     ----------
@@ -25,123 +18,230 @@ class HomologousRadial1DGeometry:
 
     def __init__(
         self,
-        v_inner,
-        v_outer,
-        v_inner_boundary,
-        v_outer_boundary,
-        time_explosion,
+        r_inner: u.Quantity,
+        r_outer: u.Quantity,
+        v_inner: u.Quantity,
+        v_outer: u.Quantity,
+        r_inner_boundary: u.Quantity,
+        r_outer_boundary: u.Quantity,
+        v_inner_boundary: u.Quantity,
+        v_outer_boundary: u.Quantity,
     ):
-        self.time_explosion = time_explosion
 
-        # ensuring that the cells are continuous
+        assert np.allclose(r_inner[1:], r_outer[:-1])
         assert np.allclose(v_inner[1:], v_outer[:-1])
 
         assert "velocity" in v_inner.unit.physical_type
         assert "velocity" in v_outer.unit.physical_type
+        assert "length" in r_inner.unit.physical_type
+        assert "length" in r_outer.unit.physical_type
 
+        self.r_inner = r_inner
+        self.r_outer = r_outer
         self.v_inner = v_inner
         self.v_outer = v_outer
 
         # ensuring that the boundaries are within the simulation area
+        if r_inner_boundary is None:
+            self.r_inner_boundary = self.r_inner[0]
+        elif r_inner_boundary < 0:
+            warnings.warn(
+                "r_inner_boundary < 0, assuming default value",
+                DeprecationWarning,
+            )
+            self.r_inner_boundary = self.r_inner[0]
+        else:
+            self.r_inner_boundary = r_inner_boundary
 
+        if r_outer_boundary is None:
+            self.r_outer_boundary = self.r_outer[-1]
+        elif r_outer_boundary < 0:
+            warnings.warn(
+                "r_outer_boundary < 0, assuming default value",
+                DeprecationWarning,
+            )
+            self.r_outer_boundary = self.r_outer[-1]
+        else:
+            self.r_outer_boundary = r_outer_boundary
+
+        assert self.r_inner_boundary < self.r_outer_boundary
+        if self.r_inner_boundary < self.r_inner[0]:
+            warnings.warn(
+                "Requesting inner boundary below inner shell. Extrapolating the inner cell"
+            )
+
+        if self.r_outer_boundary > self.r_outer[-1]:
+            warnings.warn(
+                "Requesting outer boundary above outer shell. Extrapolating the outer cell"
+            )
+
+        # Set velocity boundaries directly during construction so explicitly
+        # supplied radius boundaries remain authoritative.
         if v_inner_boundary is None:
-            self.v_inner_boundary = self.v_inner[0]
+            self._v_inner_boundary = self.v_inner[0]
         elif v_inner_boundary < 0:
             warnings.warn(
                 "v_inner_boundary < 0, assuming default value",
                 DeprecationWarning,
             )
-            self.v_inner_boundary = self.v_inner[0]
+            self._v_inner_boundary = self.v_inner[0]
         else:
-            self.v_inner_boundary = v_inner_boundary
+            self._v_inner_boundary = v_inner_boundary
 
         if v_outer_boundary is None:
-            self.v_outer_boundary = self.v_outer[-1]
+            self._v_outer_boundary = self.v_outer[-1]
         elif v_outer_boundary < 0:
             warnings.warn(
                 "v_outer_boundary < 0, assuming default value",
                 DeprecationWarning,
             )
-            self.v_outer_boundary = self.v_outer[-1]
+            self._v_outer_boundary = self.v_outer[-1]
         else:
-            self.v_outer_boundary = v_outer_boundary
+            self._v_outer_boundary = v_outer_boundary
 
-        assert self.v_inner_boundary < self.v_outer_boundary
-        if self.v_inner_boundary < self.v_inner[0]:
-            warnings.warn(
-                "Requesting inner boundary below inner shell. Extrapolating the inner cell"
-            )
+    def _radius_at_velocity(self, velocity: u.Quantity) -> u.Quantity:
+        """Return the radius corresponding to a velocity boundary.
 
-        if self.v_outer_boundary > self.v_outer[-1]:
-            warnings.warn(
-                "Requesting inner boundary below inner shell. Extrapolating the inner cell"
-            )
+        The velocity profile is piecewise linear in radius. Boundary updates
+        are therefore mapped through the first or last cell when they
+        extrapolate outside the model, matching the geometry's existing
+        boundary-extrapolation behavior.
+        """
+        velocity_boundaries = np.concatenate(
+            (self.v_inner[:1], self.v_outer)
+        )
+        radius_boundaries = np.concatenate(
+            (self.r_inner[:1], self.r_outer)
+        )
+
+        if velocity_boundaries[0] <= velocity_boundaries[-1]:
+            boundary_index = np.searchsorted(
+                velocity_boundaries, velocity, side="right"
+            ) - 1
+        else:
+            boundary_index = len(velocity_boundaries) - np.searchsorted(
+                velocity_boundaries[::-1], velocity, side="left"
+            ) - 1
+
+        boundary_index = np.clip(
+            boundary_index, 0, len(velocity_boundaries) - 2
+        )
+        velocity_gradient = (
+            velocity_boundaries[boundary_index + 1]
+            - velocity_boundaries[boundary_index]
+        ) / (
+            radius_boundaries[boundary_index + 1]
+            - radius_boundaries[boundary_index]
+        )
+        return radius_boundaries[boundary_index] + (
+            velocity - velocity_boundaries[boundary_index]
+        ) / velocity_gradient
+
+    @property
+    def v_inner_boundary(self) -> u.Quantity:
+        """Velocity at the active inner boundary."""
+        return self._v_inner_boundary
+
+    @v_inner_boundary.setter
+    def v_inner_boundary(self, velocity: u.Quantity) -> None:
+        self._v_inner_boundary = velocity
+        self.r_inner_boundary = self._radius_at_velocity(velocity)
+
+    @property
+    def v_outer_boundary(self) -> u.Quantity:
+        """Velocity at the active outer boundary."""
+        return self._v_outer_boundary
+
+    @v_outer_boundary.setter
+    def v_outer_boundary(self, velocity: u.Quantity) -> None:
+        self._v_outer_boundary = velocity
+        self.r_outer_boundary = self._radius_at_velocity(velocity)
 
     @property
     def v_middle(self):
+        """Velocity at the middle of each shell."""
         return (self.v_inner + self.v_outer) / 2.0
 
     @property
     def v_middle_active(self):
+        """Velocity at the middle of each active shell."""
         return (self.v_inner_active + self.v_outer_active) / 2.0
 
     @property
-    def v_inner_boundary_idx(self):
+    def r_middle(self):
+        """Radius at the middle of each shell."""
+        return (self.r_inner + self.r_outer) / 2.0
+
+    @property
+    def r_middle_active(self):
+        """Radius at the middle of each active shell."""
+        return (self.r_inner_active + self.r_outer_active) / 2.0
+
+    @property
+    def r_inner_boundary_index(self):
+        """Index of the shell at the active inner boundary."""
         return np.clip(
-            np.searchsorted(self.v_inner, self.v_inner_boundary, side="right")
+            np.searchsorted(self.r_inner, self.r_inner_boundary, side="right")
             - 1,
             0,
             None,
         )
 
     @property
-    def v_outer_boundary_idx(self):
+    def r_outer_boundary_index(self):
+        """Index of the shell at the active outer boundary."""
         return np.clip(
-            np.searchsorted(self.v_outer, self.v_outer_boundary, side="left")
+            np.searchsorted(self.r_outer, self.r_outer_boundary, side="left")
             + 1,
             None,
-            len(self.v_outer),
+            len(self.r_outer),
         )
 
     @property
+    def v_inner_boundary_idx(self):
+        """Index of the active inner velocity boundary."""
+        return self.r_inner_boundary_index
+
+    @property
+    def v_outer_boundary_idx(self):
+        """Index of the active outer velocity boundary."""
+        return self.r_outer_boundary_index
+
+    @property
+    def r_inner_active(self):
+        r_inner_active = self.r_inner[
+            self.r_inner_boundary_index : self.r_outer_boundary_index
+        ].copy()
+        r_inner_active[0] = self.r_inner_boundary
+        return r_inner_active
+
+    @property
+    def r_outer_active(self):
+        r_outer_active = self.r_outer[
+            self.r_inner_boundary_index : self.r_outer_boundary_index
+        ].copy()
+        r_outer_active[-1] = self.r_outer_boundary
+        return r_outer_active
+
+    @property
     def v_inner_active(self):
+        """Inner velocities for the active shells."""
+        # Indices should be the same as those used for r boundary index
         v_inner_active = self.v_inner[
-            self.v_inner_boundary_idx : self.v_outer_boundary_idx
+            self.r_inner_boundary_index : self.r_outer_boundary_index
         ].copy()
         v_inner_active[0] = self.v_inner_boundary
         return v_inner_active
 
     @property
     def v_outer_active(self):
+        """Outer velocities for the active shells."""
+        # Indices should be the same as those used for r boundary index
         v_outer_active = self.v_outer[
-            self.v_inner_boundary_idx : self.v_outer_boundary_idx
+            self.r_inner_boundary_index : self.r_outer_boundary_index
         ].copy()
         v_outer_active[-1] = self.v_outer_boundary
         return v_outer_active
-
-    @property
-    def r_inner(self):
-        return (self.v_inner * self.time_explosion).cgs
-
-    @property
-    def r_inner_active(self):
-        return (self.v_inner_active * self.time_explosion).cgs
-
-    @property
-    def r_outer(self):
-        return (self.v_outer * self.time_explosion).cgs
-
-    @property
-    def r_outer_active(self):
-        return (self.v_outer_active * self.time_explosion).cgs
-
-    @property
-    def r_middle(self):
-        return (self.v_middle * self.time_explosion).cgs
-
-    @property
-    def r_middle_active(self):
-        return (self.v_middle_active * self.time_explosion).cgs
 
     @property
     def volume(self):
@@ -156,6 +256,45 @@ class HomologousRadial1DGeometry:
             * np.pi
             * (self.r_outer_active**3 - self.r_inner_active**3)
         )
+
+    @property
+    def velocity_gradient(self):
+        """Velocity gradient in each cell"""
+        return (
+            (self.v_outer - self.v_inner)
+            / (self.r_outer - self.r_inner)
+        )
+
+    @property
+    def velocity_gradient_active(self):
+        """Velocity gradient in each cell"""
+        return (
+            (self.v_outer_active - self.v_inner_active)
+            / (self.r_outer_active - self.r_inner_active)
+        )
+
+    def get_velocity(self, r: u.Quantity, shell_id: int) -> u.Quantity:
+        """Calculate the velocity at a given radius within a shell.
+
+        Assume a piece-wise linear velocity with radius.
+
+        Parameters
+        ----------
+        r : astropy.units.quantity.Quantity
+            Radius at which to calculate the velocity
+        shell_id : int
+            Shell index
+
+        Returns
+        -------
+        astropy.units.quantity.Quantity
+            Velocity at radius r within shell shell_id
+
+        Examples
+        --------
+            geometry.get_velocity(rpacket.r, rpacket.current_shell_id)
+        """
+        return self.v_inner[shell_id] + self.velocity_gradient[shell_id] * (r - self.r_inner[shell_id])
 
     @property
     def no_of_shells(self):
@@ -187,14 +326,18 @@ numba_geometry_spec = [
     ("r_outer", float64[:]),
     ("v_inner", float64[:]),
     ("v_outer", float64[:]),
+    ("velocity_gradient", float64[:]),
     ("time_explosion", float64),
+    ("homologous", boolean),
     ("volume", float64[:]),
 ]
 
 
 @jitclass(numba_geometry_spec)
 class NumbaRadial1DGeometry:
-    def __init__(self, r_inner, r_outer, v_inner, v_outer):
+    def __init__(
+        self, r_inner, r_outer, v_inner, v_outer, homologous=False
+    ):
         """
         Radial 1D Geometry for the Numba mode
 
@@ -204,32 +347,40 @@ class NumbaRadial1DGeometry:
         r_outer : numpy.ndarray
         v_inner : numpy.ndarray
         v_outer : numpy.ndarray
+        velocity_gradient : numpy.ndarray
+        homologous : bool
         volume : numpy.ndarray
         """
         self.r_inner = r_inner
         self.r_outer = r_outer
         self.v_inner = v_inner
         self.v_outer = v_outer
-        self.time_explosion = (
-            self.r_outer[0] / self.v_outer[0]
-        )  # avoids potential division by zero if v_inner is 0
+        self.velocity_gradient = (self.v_outer - self.v_inner) / (self.r_outer - self.r_inner)
+        self.homologous = homologous
+        self.time_explosion = 0.0
+        if self.homologous:
+            self.time_explosion = self.r_outer[0] / self.v_outer[0]
         self.volume = (4 / 3) * np.pi * (self.r_outer**3 - self.r_inner**3)
 
     def get_velocity(self, r: float, shell_id: int) -> float:
-        """
-        Calculate homologous velocity at a radius.
+        """Calculate the velocity at a given radius within a shell.
+
+        Assume a piece-wise linear velocity with radius.
 
         Parameters
         ----------
         r : float
-            Radius at which to calculate the velocity [cm].
+            Radius at which to calculate the velocity
         shell_id : int
-            Shell index. Included for interface parity with non-homologous
-            geometry; not used for homologous expansion.
+            Shell index
 
         Returns
         -------
         float
-            Homologous velocity at radius r [cm / s].
+            Velocity at radius r within shell shell_id
         """
-        return r / self.time_explosion
+        if self.homologous:
+            return r / self.time_explosion
+        return self.v_inner[shell_id] + self.velocity_gradient[shell_id] * (
+            r - self.r_inner[shell_id]
+        )
