+++ tardis/spectrum/formal_integral/tests/test_cuda_formal_integral.py
@@ -76,6 +76,7 @@ def formal_integral_geometry(request):
         r[1:],
         r[:-1] * c.c.cgs.value,
         r[1:] * c.c.cgs.value,
+        homologous=True,
     )
     return geometry
