+++ tardis/transport/montecarlo/modes/nonhomologous/packet_propagation.py
@@ -4,8 +4,8 @@ import numpy as np
 from numba import njit
 
 from tardis import constants as const
-from tardis.model.geometry.radial1d_nonhomologous import (
-    NumbaNonhomologousRadial1DGeometry,
+from tardis.model.geometry.radial1d import (
+    NumbaRadial1DGeometry,
 )
 from tardis.opacities.opacities import chi_electron_calculator
 from tardis.opacities.opacity_state_numba import OpacityStateNumba
@@ -53,7 +53,7 @@ C_SPEED_OF_LIGHT = const.c.to("cm/s").value
 @njit
 def packet_propagation(
     r_packet: RPacket,
-    geometry: NumbaNonhomologousRadial1DGeometry,
+    geometry: NumbaRadial1DGeometry,
     time_explosion: float,
     opacity_state: OpacityStateNumba,
     estimators_bulk: EstimatorsBulk,
@@ -72,7 +72,7 @@ def packet_propagation(
     ----------
     r_packet : RPacket
         The radiative packet to transport through the ejecta.
-    geometry : NumbaNonhomologousRadial1DGeometry
+    geometry : NumbaRadial1DGeometry
         The spherically symmetric geometry of the supernova ejecta.
     time_explosion : float
         Time since explosion in seconds. Accepted for interface parity with
@@ -261,7 +261,7 @@ def packet_propagation(
 @njit
 def set_packet_props_partial_relativity(
     r_packet: RPacket,
-    geometry: NumbaNonhomologousRadial1DGeometry
+    geometry: NumbaRadial1DGeometry
 ) -> None:
     """
     Set packet properties using partial relativistic corrections.
@@ -274,7 +274,7 @@ def set_packet_props_partial_relativity(
     r_packet
         The radiative packet whose properties will be modified.
     geometry
-        NumbaNonhomologousRadial1DGeometry object
+        NumbaRadial1DGeometry object
 
     Returns
     -------
