+++ tardis/transport/montecarlo/packets/tests/test_packet.py
@@ -13,8 +13,8 @@ import tardis.transport.montecarlo.modes.homologous_rad_packet_transport as r_pa
 import tardis.transport.montecarlo.packets.radiative_packet as radiative_packet
 import tardis.transport.montecarlo.utils as utils
 from tardis import constants as const
-from tardis.model.geometry.radial1d_nonhomologous import (
-    NumbaNonhomologousRadial1DGeometry,
+from tardis.model.geometry.radial1d import (
+    NumbaRadial1DGeometry,
 )
 from tardis.transport.montecarlo.estimators.radfield_estimator_calcs import (
     update_estimators_line,
@@ -34,11 +34,12 @@ def geometry():
     time_explosion = 5.2e7
     r_inner = np.array([6.912e14, 8.64e14], dtype=np.float64)
     r_outer = np.array([8.64e14, 1.0368e15], dtype=np.float64)
-    return NumbaNonhomologousRadial1DGeometry(
+    return NumbaRadial1DGeometry(
         r_inner=r_inner,
         r_outer=r_outer,
         v_inner=r_inner / time_explosion,
         v_outer=r_outer / time_explosion,
+        homologous=True,
     )
